requirements.txt
pyserial
asyncio
websockets